+++ frontend/catalyst/tracing/type_signatures.py
@@ -77,7 +77,8 @@ def filter_static_args(args, static_argnums):
     return tuple(args[idx] for idx in range(len(args)) if idx not in static_argnums)
 
 
-def split_static_args(args, static_argnums):
+# TODO: remove pragma in part 2
+def split_static_args(args, static_argnums):  # pragma: nocover
     """Split arguments into static and dynamic values using the provided index list.
 
     Args:
@@ -120,7 +121,8 @@ def merge_static_args(signature, args, static_argnums):
     return tuple(merged_sig)
 
 
-def get_decomposed_signature(args, static_argnums):
+# TODO: remove pragma in part 2
+def get_decomposed_signature(args, static_argnums):  # pragma: nocover
     """Decompose function arguments into dynamic and static arguments, where the dynamic arguments
     are further processed into abstract values and PyTree metadata. All values returned by this
     function are hashable.
